requirements.txt
streamlit
pandas
numpy
plotly
scikit-learn
Pillow
requests
openai
matplotlib